realestate/requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
click==6.7
fake-useragent==0.1.10
Flask==1.0.2
get==1.0.3
gunicorn==19.9.0
idna==2.7
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mongoengine==0.15.3
numpy==1.15.1
pandas==0.23.4
pipenv==2018.7.1
post==1.0.2
pprint==0.1
public==1.0.3
pymongo==3.7.0
python-dateutil==2.7.3
pytz==2018.5
query-string==1.0.2
request==1.0.2
requests==2.19.1
selenium==3.14.0
six==1.11.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
sklearn
pillow
